[FIX] Folder structure and targets
