--- NetworkeengFramework/README.md
@@ -1,3 +1,0 @@
-# NetworkeengFramework
-
-A description of this package.
