--- networkeeng/.swift-format.json
@@ -1,54 +1,0 @@
-{
-  "fileScopedDeclarationPrivacy" : {
-    "accessLevel" : "private"
-  },
-  "indentation" : {
-    "spaces" : 2
-  },
-  "indentConditionalCompilationBlocks" : true,
-  "indentSwitchCaseLabels" : false,
-  "lineBreakAroundMultilineExpressionChainComponents" : false,
-  "lineBreakBeforeControlFlowKeywords" : false,
-  "lineBreakBeforeEachArgument" : false,
-  "lineBreakBeforeEachGenericRequirement" : false,
-  "lineLength" : 80,
-  "maximumBlankLines" : 1,
-  "prioritizeKeepingFunctionOutputTogether" : false,
-  "respectsExistingLineBreaks" : true,
-  "rules" : {
-    "AllPublicDeclarationsHaveDocumentation" : true,
-    "AlwaysUseLowerCamelCase" : true,
-    "AmbiguousTrailingClosureOverload" : true,
-    "BeginDocumentationCommentWithOneLineSummary" : false,
-    "DoNotUseSemicolons" : true,
-    "DontRepeatTypeInStaticProperties" : true,
-    "FileScopedDeclarationPrivacy" : true,
-    "FullyIndirectEnum" : true,
-    "GroupNumericLiterals" : true,
-    "IdentifiersMustBeASCII" : true,
-    "NeverForceUnwrap" : false,
-    "NeverUseForceTry" : false,
-    "NeverUseImplicitlyUnwrappedOptionals" : false,
-    "NoAccessLevelOnExtensionDeclaration" : true,
-    "NoBlockComments" : true,
-    "NoCasesWithOnlyFallthrough" : true,
-    "NoEmptyTrailingClosureParentheses" : true,
-    "NoLabelsInCasePatterns" : true,
-    "NoLeadingUnderscores" : false,
-    "NoParensAroundConditions" : true,
-    "NoVoidReturnOnFunctionSignature" : true,
-    "OneCasePerLine" : true,
-    "OneVariableDeclarationPerLine" : true,
-    "OnlyOneTrailingClosureArgument" : true,
-    "OrderedImports" : true,
-    "ReturnVoidInsteadOfEmptyTuple" : true,
-    "UseLetInEveryBoundCaseVariable" : true,
-    "UseShorthandTypeNames" : true,
-    "UseSingleLinePropertyGetter" : true,
-    "UseSynthesizedInitializer" : true,
-    "UseTripleSlashForDocumentationComments" : true,
-    "ValidateDocumentationComments" : false
-  },
-  "tabWidth" : 2,
-  "version" : 1
-}
