+++ networkeeng/TestApplication/Assets.xcassets/AppIcon.appiconset/Contents.json
