+++ networkeeng/TestApplication/TestApplicationApp.swift
@@ -1,0 +1,17 @@
+//
+//  TestApplicationApp.swift
+//  TestApplication
+//
+//  Created by Mark Angelo Hernandez on 19/08/2021.
+//
+
+import SwiftUI
+
+@main
+struct TestApplicationApp: App {
+  var body: some Scene {
+    WindowGroup {
+      ContentView()
+    }
+  }
+}
