--- networkeeng/demo/DemoApp.swift
@@ -1,17 +1,0 @@
-//
-//  DemoApp.swift
-//  demo
-//
-//  Created by Mark Angelo Hernandez on 27/06/2021.
-//
-
-import SwiftUI
-
-@main
-struct DemoApp: App {
-  var body: some Scene {
-    WindowGroup {
-      ContentView()
-    }
-  }
-}
