+++ networkeeng/networkeeng.xcodeproj/xcshareddata/xcschemes/NetworkeengTests.xcscheme
@@ -5,22 +5,6 @@
    <BuildAction
       parallelizeBuildables = "YES"
       buildImplicitDependencies = "YES">
-      <BuildActionEntries>
-         <BuildActionEntry
-            buildForTesting = "YES"
-            buildForRunning = "YES"
-            buildForProfiling = "YES"
-            buildForArchiving = "YES"
-            buildForAnalyzing = "YES">
-            <BuildableReference
-               BuildableIdentifier = "primary"
-               BlueprintIdentifier = "Networkeeng"
-               BuildableName = "Networkeeng"
-               BlueprintName = "Networkeeng"
-               ReferencedContainer = "container:">
-            </BuildableReference>
-         </BuildActionEntry>
-      </BuildActionEntries>
    </BuildAction>
    <TestAction
       buildConfiguration = "Debug"
@@ -32,10 +16,10 @@
             skipped = "NO">
             <BuildableReference
                BuildableIdentifier = "primary"
-               BlueprintIdentifier = "NetworkeengTests"
-               BuildableName = "NetworkeengTests"
+               BlueprintIdentifier = "A60A688126CE281B00962757"
+               BuildableName = "NetworkeengTests.xctest"
                BlueprintName = "NetworkeengTests"
-               ReferencedContainer = "container:">
+               ReferencedContainer = "container:Networkeeng.xcodeproj">
             </BuildableReference>
          </TestableReference>
       </Testables>
@@ -57,15 +41,6 @@
       savedToolIdentifier = ""
       useCustomWorkingDirectory = "NO"
       debugDocumentVersioning = "YES">
-      <MacroExpansion>
-         <BuildableReference
-            BuildableIdentifier = "primary"
-            BlueprintIdentifier = "Networkeeng"
-            BuildableName = "Networkeeng"
-            BlueprintName = "Networkeeng"
-            ReferencedContainer = "container:">
-         </BuildableReference>
-      </MacroExpansion>
    </ProfileAction>
    <AnalyzeAction
       buildConfiguration = "Debug">
